+++ opal/opal/native.rb
@@ -49,6 +49,13 @@ class Native < BasicObject
       var prop = #{native}[#{symbol}];
 
       if (typeof(prop) === 'function') {
+        for (var arg, len = args.length, i = 0; i < len; i++) {
+          arg = args[i]
+          if (arg.$to_native) {
+            args[i] = arg.$to_native();
+          }
+        }
+
         prop = prop.apply(#{native}, #{args});
 
         if (typeof(prop) === 'object' || typeof(prop) === 'function') {